Address dimas-b review comments on catalog update validation
- Use toString() instead of toASCIIString() in RestCatalogMinIOSpecialIT
- Remove redundant file://tmp from allowed locations in AbstractIcebergCatalogViewTest;
  file:// already subsumes file://tmp via StorageLocation.isChildOf() prefix matching
- Add Builder.validateDefaultBaseLocation() to consolidate location validation
  inside CatalogEntity.Builder; remove the duplicate explicit call from
  PolarisAdminService for the new-storage-config update case
- Move validateBaseLocationAgainstAllowedList tests from BaseLocationValidationTest
  into CatalogEntityTest; delete BaseLocationValidationTest
