+++ runtime/service/src/intTest/java/org/apache/polaris/service/it/RestCatalogMinIOSpecialIT.java
@@ -217,7 +217,7 @@ public class RestCatalogMinIOSpecialIT {
             .setStorageType(StorageConfigInfo.StorageTypeEnum.S3)
             .setPathStyleAccess(pathStyleAccess)
             .setStsUnavailable(!stsEnabled)
-            .setAllowedLocations(List.of(storageBase.toASCIIString() + "/" + catalogName));
+            .setAllowedLocations(List.of(storageBase.toString() + "/" + catalogName));
 
     endpoint.ifPresent(storageConfig::setEndpoint);
     stsEndpoint.ifPresent(storageConfig::setStsEndpoint);
