+++ runtime/service/src/test/java/org/apache/polaris/service/catalog/iceberg/AbstractLocalIcebergCatalogViewTest.java
@@ -180,9 +180,7 @@ public abstract class AbstractLocalIcebergCatalogViewTest
                 .setStorageConfigurationInfo(
                     realmConfig,
                     new FileStorageConfigInfo(
-                        StorageConfigInfo.StorageTypeEnum.FILE,
-                        List.of("file://tmp", "file://", "/", "*"),
-                        null),
+                        StorageConfigInfo.StorageTypeEnum.FILE, List.of("file://", "/", "*"), null),
                     "file://tmp")
                 .build()
                 .asCatalog(serviceIdentityProvider)));
